Add BackgroundStrokeDrawable
