+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/utils/extensions/Int.kt
@@ -2,4 +2,5 @@ package io.getstream.chat.android.ui.utils.extensions
 
 import android.content.res.Resources
 
-internal fun Int.dpToPx() = (this * Resources.getSystem().displayMetrics.density).toInt()
+internal fun Int.dpToPx(): Int = dpToPxPrecise().toInt()
+internal fun Int.dpToPxPrecise(): Float = (this * Resources.getSystem().displayMetrics.density)
