+++ stream-chat-android-ui-components/src/main/res/layout/stream_item_message_plain_text.xml
@@ -19,7 +19,6 @@
         android:paddingTop="@dimen/stream_spacing_small"
         android:paddingRight="@dimen/stream_spacing_medium"
         android:paddingBottom="@dimen/stream_spacing_small"
-        android:text="@string/stream_message_deleted_label"
         android:textAppearance="@style/TextAppearance.MaterialComponents.Body2"
         android:lineHeight="16sp"
         android:gravity="start"
